+++ CHANGES.rst
@@ -28,6 +28,11 @@ Internal Changes
 
 - Make React and SemanticUI usable everywhere (:issue:`3955`)
 
+Bugfixes
+^^^^^^^^
+
+- Hide Book of Abstracts menu item if LaTeX is disabled
+
 
 ----